requirements.txt
pyautogen==0.2.28
playwright==1.41.0
requests
chromadb
sentence-transformers
pypdf
